requirements.txt
# Imagefront Framework - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
python-dotenv>=1.0.0

# Google Generative AI (for Imagen/Nano Banana - PRIMARY)
google-generativeai>=0.3.0

# Alternative: OpenAI (for DALL-E 3 and GPT-4V)
openai>=1.0.0

# Alternative: Replicate (for Stable Diffusion, Flux, etc.)
replicate>=0.20.0

# Image processing
Pillow>=10.0.0
requests>=2.31.0

# Optional: Annotation tools
# opencv-python>=4.8.0
